backend/test_db.py
import pytest
import psycopg2
import collections

# A very basic mock test to ensure test coverage runs without needing a real DB
def test_mock_db_connection(monkeypatch):
    class MockConnection:
        def __init__(self):
            self.autocommit = False
            self.closed = False
        def close(self):
            self.closed = True
    
    mock_conn = MockConnection()
    
    def mock_connect(*args, **kwargs):
        return mock_conn

    monkeypatch.setattr(psycopg2, "connect", mock_connect)
    
    conn = psycopg2.connect("dummy")
    assert conn is mock_conn
    assert not conn.closed
    conn.close()
    assert conn.closed
