backend/requirements.txt
flask==3.0.0
psycopg2-binary==2.9.9
redis==5.0.1
flask-cors==4.0.0
gunicorn==21.2.0
pytest==7.4.3
pytest-flask==1.3.0
